requirements.txt
streamlit==1.52.1
python-dotenv==1.0.1
